Crate tokio_openssl

Async TLS streams backed by OpenSSL

This library is an implementation of TLS streams using OpenSSL for negotiating the connection. Each TLS stream implements the Read and Write traits to interact and interoperate with the rest of the futures I/O ecosystem. Client connections initiated from this crate verify hostnames automatically and by default.

This crate primarily exports this ability through two extension traits, SslConnectorExt and SslAcceptorExt. These traits augment the functionality provided by the openssl crate, on which this crate is built. Configuration of TLS parameters is still primarily done through the openssl crate.

Structs

HandshakeError

The error type returned after a failed handshake.

HandshakeFuture

SslStream

A wrapper around an underlying raw stream which implements the SSL protocol.

StartHandshakeFuture

StartHandshakeFutureInner

StreamWrapper

Enums

StartedHandshake

Functions

accept

Asynchronously performs a server-side TLS handshake over the provided stream.

connect

Asynchronously performs a client-side TLS handshake over the provided stream.

cvt

handshake