Struct tokio_openssl::SslStream

pub struct SslStream<S>(pub(crate) SslStream<StreamWrapper<S>>);

An asynchronous version of openssl::ssl::SslStream.

Tuple Fields
0: SslStream<StreamWrapper<S>>
Implementations

impl<S> SslStream<S> where
S: AsyncRead + AsyncWrite,

pub fn new(ssl: Ssl, stream: S) -> Result<Self, ErrorStack>

Like SslStream::new.

pub fn poll_connect(
self: Pin<&mut Self>,
cx: &mut Context<'_>
) -> Poll<Result<(), Error>>

Like SslStream::connect.

pub async fn connect(self: Pin<&mut Self>) -> Result<(), Error>

A convenience method wrapping poll_connect.

pub fn poll_accept(
self: Pin<&mut Self>,
cx: &mut Context<'_>
) -> Poll<Result<(), Error>>

Like SslStream::accept.

pub async fn accept(self: Pin<&mut Self>) -> Result<(), Error>

A convenience method wrapping poll_accept.

pub fn poll_do_handshake(
self: Pin<&mut Self>,
cx: &mut Context<'_>
) -> Poll<Result<(), Error>>

Like SslStream::do_handshake.

pub async fn do_handshake(self: Pin<&mut Self>) -> Result<(), Error>

A convenience method wrapping poll_do_handshake.

pub fn poll_read_early_data(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
buf: &mut [u8]
) -> Poll<Result<usize, Error>>

Like SslStream::read_early_data.

pub async fn read_early_data(
self: Pin<&mut Self>,
buf: &mut [u8]
) -> Result<usize, Error>

A convenience method wrapping poll_read_early_data.

pub fn poll_write_early_data(
self: Pin<&mut Self>,
cx: &mut Context<'_>,
buf: &[u8]
) -> Poll<Result<usize, Error>>

Like SslStream::write_early_data.

pub async fn write_early_data(
self: Pin<&mut Self>,
buf: &[u8]
) -> Result<usize, Error>

A convenience method wrapping poll_write_early_data.

impl<S> SslStream<S>

pub fn ssl(&self) -> &SslRef

Returns a shared reference to the Ssl object associated with this stream.

pub fn get_ref(&self) -> &S

Returns a shared reference to the underlying stream.

pub fn get_mut(&mut self) -> &mut S

Returns a mutable reference to the underlying stream.

pub fn get_pin_mut(self: Pin<&mut Self>) -> Pin<&mut S>

Returns a pinned mutable reference to the underlying stream.

pub(crate) fn with_context<F, R>(
self: Pin<&mut Self>,
ctx: &mut Context<'_>,
f: F
) -> R where
F: FnOnce(&mut SslStream<StreamWrapper<S>>) -> R,